Move `test_plotly.rs` under `tests/unit/visualization` and rename modules for clarity and feature-specific segregation.
